Fix: mitigated double-clicks on redeem
Implemented `redeeming.value` reactive state to mitigate double-clicking on buttons.

There is a chance that Sweetalert (the popup) is allowing double-clicks to go through upon confirmation, so I have updated the dialog to request only a single click when confirming.

We may need to move the redemption confirmation dialog to a proprietary modal instead - thoughts?